requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pyserial>=3.5
websockets>=12.0
pyyaml>=6.0
